+++ packages/arb-token-bridge-ui/src/components/TransferPanel/TransferPanel.eth-weth.integration.test.tsx
@@ -1,0 +1,56 @@
+import { constants } from 'ethers';
+import { describe, it } from 'vitest';
+
+import {
+  ethWethCases,
+  expectTokenButtonSymbol,
+  expectTokenPanelSymbol,
+  nonConnectedDestinationAddress,
+  renderTransferPanel,
+  setupTransferPanelLifiIntegrationSuite,
+} from './TransferPanel.integration.helpers';
+
+describe.sequential('TransferPanel LiFi Integration - ETH/WETH Override', () => {
+  setupTransferPanelLifiIntegrationSuite();
+
+  it.each(ethWethCases)(
+    'renders source $expectedSourceSymbol and destination $expectedDestinationSymbol for ETH/WETH override: $sourceChain -> $destinationChain',
+    async ({
+      sourceChain,
+      destinationChain,
+      expectedSourceSymbol,
+      expectedDestinationSymbol,
+      expectedSourcePanelSymbols,
+      expectedDestinationPanelSymbols,
+    }) => {
+      renderTransferPanel({
+        sourceChain,
+        destinationChain,
+        destinationToken: constants.AddressZero,
+        destinationAddress: nonConnectedDestinationAddress,
+      });
+
+      await expectTokenButtonSymbol({
+        isDestination: false,
+        symbol: expectedSourceSymbol,
+      });
+      await expectTokenButtonSymbol({
+        isDestination: true,
+        symbol: expectedDestinationSymbol,
+      });
+
+      if (expectedSourcePanelSymbols) {
+        await expectTokenPanelSymbol({
+          isDestination: false,
+          symbolsToContain: expectedSourcePanelSymbols,
+        });
+      }
+      if (expectedDestinationPanelSymbols) {
+        await expectTokenPanelSymbol({
+          isDestination: true,
+          symbolsToContain: expectedDestinationPanelSymbols,
+        });
+      }
+    },
+  );
+});
