+++ packages/arb-token-bridge-ui/src/components/TransferPanel/TransferPanel.usdc.integration.test.tsx
@@ -1,0 +1,64 @@
+import { describe, it } from 'vitest';
+
+import {
+  expectTokenButtonSymbol,
+  expectTokenPanelSymbol,
+  getUsdcSourceToken,
+  nonConnectedDestinationAddress,
+  renderTransferPanel,
+  setupTransferPanelLifiIntegrationSuite,
+  usdcAddressByChain,
+  usdcCases,
+} from './TransferPanel.integration.helpers';
+
+describe.sequential('TransferPanel LiFi Integration - USDC', () => {
+  setupTransferPanelLifiIntegrationSuite();
+
+  it.each(usdcCases)(
+    'renders source $expectedSourceSymbol and destination $expectedDestinationSymbol for USDC transfer: $sourceChain -> $destinationChain',
+    async ({
+      sourceChain,
+      destinationChain,
+      expectedSourceSymbol,
+      expectedDestinationSymbol,
+      expectedSourcePanelSymbols,
+      expectedDestinationPanelSymbols,
+    }) => {
+      const sourceUsdcAddress = usdcAddressByChain[sourceChain];
+      const sourceUsdcToken = getUsdcSourceToken(sourceChain);
+
+      renderTransferPanel({
+        sourceChain,
+        destinationChain,
+        token: sourceUsdcAddress,
+        destinationToken: sourceUsdcAddress,
+        bridgeTokens: {
+          [sourceUsdcAddress.toLowerCase()]: sourceUsdcToken,
+        },
+        destinationAddress: nonConnectedDestinationAddress,
+      });
+
+      await expectTokenButtonSymbol({
+        isDestination: false,
+        symbol: expectedSourceSymbol,
+      });
+      await expectTokenButtonSymbol({
+        isDestination: true,
+        symbol: expectedDestinationSymbol,
+      });
+
+      if (expectedSourcePanelSymbols) {
+        await expectTokenPanelSymbol({
+          isDestination: false,
+          symbolsToContain: expectedSourcePanelSymbols,
+        });
+      }
+      if (expectedDestinationPanelSymbols) {
+        await expectTokenPanelSymbol({
+          isDestination: true,
+          symbolsToContain: expectedDestinationPanelSymbols,
+        });
+      }
+    },
+  );
+});
